refactor(vm): decompose vm_run.c into responsibility modules
vm_run.c was the file every VM fix touches, and the v1.3.5 codebase-health
scope calls for splitting it. It carried five unrelated responsibilities in
one translation unit, and — because the interpreter has two dispatch
mechanisms (computed-goto threading on GCC/Clang, a `switch` loop elsewhere) —
it carried most opcode bodies TWICE, once per mechanism, with nothing keeping
the two copies in step.

The split is by responsibility, matching how vm_core.c / vm_native.c /
vm_parser.c / vm_macro.c are already separated:

  vm_run.c        the dispatch loop: fetch, decode, route
  vm_ops.c        numeric-operand guards, comparison, pair, vector and
                  operand-stack opcode bodies
  vm_frame.c      upvalue access, closure construction, the return sequence
  vm_control.c    continuation capture/resume, dynamic-wind rerooting, the
                  exception-handler stack
  vm_limits.c     the runaway-instruction guard and timeout checkpoint
  vm_lifecycle.c  vm_create()/vm_free() and the test-program assembler

Behaviour-preserving, and mechanically so: every moved body is the ORIGINAL
text, and every opcode body that is now shared by both dispatch mechanisms was
verified byte-identical (whitespace-insensitive) between the two copies before
being unified. The bodies whose two copies are NOT equivalent today stay
inline in vm_run.c rather than being unified, because unifying them would
change behaviour: the arithmetic opcodes (only the threaded path records
reverse-mode AD tape nodes), OP_GET_LOCAL/OP_SET_LOCAL, OP_STR_REF/OP_STR_LEN,
OP_PRINT, OP_CALL/OP_TAIL_CALL, OP_CALLCC, OP_WIND_PUSH, OP_CONST and
OP_GLOBAL_MARK.

Error signalling is unchanged: a body reports failure by setting vm->error (or
vm->halted) and returning, and the caller's next dispatch step observes the
flag and leaves the loop — exactly what the `goto vm_exit` / `break` it
replaced did.

The continuation-resume sequence had six copies (OP_CALL, OP_TAIL_CALL and
OP_INVOKE_CC, each once per dispatch mechanism) and is now one function.

Build wiring: the five new files are unity-build components like their
siblings, #included by eshkol_vm.c (opcode modules before vm_run.c, which
calls them; after vm_native.c and vm_region_evac.c, which they call), listed
in ESHKOL_VM_CORE_COMPONENT_SRC, and required by vm_source_boundary_test so
the classification cannot silently lapse. The ICC architecture model and the
VM layout docs describe the new boundaries.
